requirements.txt
langchain
chromadb
tiktoken
chainlit
openai
python-dotenv
langchain-community
PyPDF2
pillow
pymupdf